Fixed: no such file or directory

+++ index.js
@@ -123,7 +123,9 @@ const previousPath = process.cwd();
 process.chdir(tempDir);
 execSync(`gzip archtemp.tar`);
 // Move the Unity Package (.gz) to the original path
-fs.rmSync(unityPackagePath);
+if (fs.existsSync(unityPackagePath)) {
+    fs.rmSync(unityPackagePath);
+}
 fs.renameSync(`${tempDir}/archtemp.tar.gz`, unityPackagePath);
 process.chdir(previousPath);
 // endregion
